Update regex for empty nodes for also excluding img
Some websites using closing `</img>` nodes which is not defined in HTML. In most cases, this results in `<img src="..."></img>` which is currently detected as empty node and therefore removed by graby. So all images from those domains can currently not be fetched by wallabag.

This will fix lot of issues in wallabagger for missing images.